Robust CI Test Job
Test JSDoc template as it will be used by PlayCanvas Engine repository.

This is the full end to end test as described in the readme. Once the "docs" command of the engine repo is confirmed working, this will modify the configuration file to use the updated code in the pull request or commit, then build PlayCanvas API Reference.

Test will gather artifacts to validate the operation of proposed changes.  See the test run history for details.