Update publish.js
Further compatibility fixes to remove usage of taffydb 'first' and 'order'.

The order function does an locale aware alphabetic sort (does not move capitol letters to the top).

This uncovered a bug in the alphaSort function but it has not been replaced to ensure the same output is generated as before.  This can be seen in a handful of documentation pages where items are out of order because they have a capitol letter.